Seção 9: DTOs e Data Annotations

+++ Back/src/ProEventos.API/Controllers/EventosController.cs
@@ -1,12 +1,9 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using ProEventos.Application.Dtos;
 using ProEventos.Application.Contratos;
-using ProEventos.Domain;
-using ProEventos.Persistence.Contextos;
 
 namespace ProEventos.API.Controllers
 {
@@ -28,7 +25,7 @@ namespace ProEventos.API.Controllers
            try
            {
              var eventos = await _eventoService.GetAllEventosAsync(true);
-             if  (eventos == null) return NotFound("Nenhum evento encontrado.");
+             if  (eventos == null) NoContent();
 
              return Ok(eventos);
            }
@@ -45,7 +42,7 @@ namespace ProEventos.API.Controllers
            try
            {
              var evento = await _eventoService.GetEventoByIdAsync(id, true);
-             if  (evento == null) return NotFound("Evento por Id não encontrado.");
+             if  (evento == null) return NoContent();
 
              return Ok(evento);
            }
@@ -62,7 +59,7 @@ namespace ProEventos.API.Controllers
            try
            {
              var evento = await _eventoService.GetAllEventosByTemaAsync(tema, true);
-             if  (evento == null) return NotFound("Eventos por tema não encontrados.");
+             if  (evento == null) return NoContent();
 
              return Ok(evento);
            }
@@ -74,12 +71,12 @@ namespace ProEventos.API.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> Post(Evento model)
+        public async Task<IActionResult> Post(EventoDto model)
         {
             try
            {
              var evento = await _eventoService.AddEventos(model);
-             if  (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+             if  (evento == null) return NoContent();
 
              return Ok(evento);
            }
@@ -91,12 +88,12 @@ namespace ProEventos.API.Controllers
         }
 
          [HttpPut("{id}")]
-        public async Task<IActionResult> Put(int id, Evento model)
+        public async Task<IActionResult> Put(int id, EventoDto model)
         {
             try
             {
                 var evento = await _eventoService.UpdateEvento(id, model);
-                if (evento == null) return BadRequest("Erro ao tentar adicionar evento.");
+                if (evento == null) return NoContent();
 
                 return Ok(evento);
             }
@@ -112,9 +109,12 @@ namespace ProEventos.API.Controllers
         {
             try
             {
+                var evento = await _eventoService.GetEventoByIdAsync(id, true);
+                if  (evento == null) return NoContent();
+
                 return await _eventoService.DeleteEvento(id) ? 
                        Ok("Deletado") : 
-                       BadRequest("Evento não deletado");
+                       throw new Exception("Ocorreu um problema não específico ao tentar deletar o Evento.");
             }
             catch (Exception ex)
             {
